Refactoring the index generator.
The HTML-aware part of the generator has been extracted into a newly created
Html class. This class has only static methods, which generate HTML code. A few
examples: link, enclose, post_summary, doc_header. Neither of these methods
know anything about posts, maildb-s, postsets or really anything other than
HTML.

The code that remained in the Generator class deals with the posts and the mail
database. The code were divided into smaller methods that generate HTML-code
for well-defined entities: a post summary, a thread, a section, or the whole
database. ("Post summary" is the one-line summary of the post that includes the
author, the subject, the tags, the index and the date. These were printed in
the index before, of course, only the name "post summary" is new.)

The cleaner code is a great benefit, but the greatest benefit (and the main
reason of this change) is the testability of the Generator. All methods of the
Generator are tested now (except for post_html, on which we plan to work in the
near future).

The generator is a bit slower. I calculated the means of 10 executions of the
thread calculation and index generation on the UMS Heap, both with the old and
the new Generator. Using the new one, it takes 0.485 seconds to run; using the
old one, it takes 0.435 seconds. Probably the memory usage is slightly higher,
as well. The reason is that now we create large strings in the memory, while
the old one used to write almost everything on the output stream. But this
tradeoff is worth it, I think. (I tried using output streams in the new
version, but I gave up quickly, as it would have been much more complicated to
implement: the function call structure would be forced to follow the structure
of the HTML to be generated.)

Changes that introduce incompatibily:
- heapcustom: now datefun should return the date in parentheses if the user
  wishes to see them in parentheses
- heapcustom: gen_index_html should be renamed to gen_index
- heapmanip: Generator.index_html has been renamed to Generator.index, thus the
  body of heapcustom.gen_index should be changed
- heapia: heapia option 'gen_index_html' has been renamed to 'gen_index'