Docstrings improved, bugs fixed.
The documentation of the "not real types" (i.e. types that exist only on
the level of our abstract thinking) and the documentation of the
heapmanip.SectionOptions, heapmanip.GeneratorOptions and
heapcustomlib.DateOptions types has been improved and cleaned up.

A bug in heapia has been fixed, and the default implementation of the
heapia.gen_index has been removed.

A bug in heapcustom-csabahoch was fixed.