Sections are described by Section objects.
A class called heapmanip.Section has been created, which describes a
section: its title, posts and whether it should be printed flatly. Other
options may be added later; the class was designed so that further
addition will require modification of any existing code.

This pattern seems to be good for representing options, as well, so the
GeneratorOptions will be refactored to use this pattern. The options of
heapia will be also refactored in this style.

The class heapmanip.Index is to be created; that should follow the same
pattern as Section does.

A constant was introduced in heaplib:

    class NOT_SET: pass

NOT_SET is an empty class, but it is used as a constant. This peculiar
way of defining a constant was used to ensure that this constant will
not be equal to anything but itself. Less ugly solutions are welcome.