heapia: implementing event handling and "features"
Heapia has been refactored and improved. Now it has two customization
levels: features for basic users and event handling for advanced users.
Features are implemented using event handling.

Features
--------

Features can be turned on and off with heapia commands 'on' and 'off',
just like this:

    >>> on('gen_incices')
    >>> off('gen_incices')

The 'gen_indices' feature will regenerate the indices (HTML files that
contain an index of the posts) when a mail database changed after a
commands. Other features:

* gen_indices - automatically regenerates the indices
* gen_posts - automatically regenerates the posts
* save - automatically saves the database
* timer - times the commands
* touched_post_printer - prints the list of posts that has been changed
  by the last command

Every feature has a shortcut, so for example it is enough to type
"on('gi')" instead of "on('gen_incices')".

Event handling
--------------

Now heapia uses an event handling model instead of ad hoc event
handling. heapia.listeners is a list that contains objects or functions,
which are called listeners or event handlers. They are called each time
an event happens in heapia. An event is raised when a command is called,
a command is finished, or anything really that we want to use as an
event.

All features are implemented as event handlers. The event handlers can
depend on each other: e.g. the event handler of most features depend on
the TouchedPostPrinterListener event printer. Thus this event printer is
present in the heapia.listeners list by default.

heapmanip.MailDB now has a notification service, as well, which works
very similarly to the event handling implemented in heapia. MailDB
stores a list of listeners that are called when the MailDB is touched.
(If a post was touched, the listener will get a reference to the touched
post. This way any object that is a listener of the MailDB can track
which posts are touched and when.)