Doc/Arch: general intro and dependencies added
An introduction and the "Module dependencies" subsection has been added to the
Architecture page of the documentation. The latter also contains a png picture
generated from a dot file; both has been added to the repository. The format of
the Architecture page probably will not change too much.

The Renamings item in the todo has been modified.

Minor corrections on heapmanip_explained were made.