hklib.Post: bug fixed
The implementation of hklib.Post relies on the invariant that all attributes of
the Post._header are always defined. This invariant was not true after deleting
a post, which could cause problems e.g. after using the hkshell.d() function to
delete the post and then the event printer to print the set of modified post.
Now this invariant has been fixed by modifying the Post.delete function.