hkshell: introducing another decorator
The hkshell.perform_operation function has been replaced by the
hkshell.postset_operation decorator, which is more convenient to use. Although
it may be a very little confusing, because the decorator changes the decorated
function's parameter list. But this is all documented in postset_operation's
docstring and in the HTML user documentation of hkshell.

The docstrings of hkshell have been updated to the new format. There are many
missing docstrings though.