hkshell: new parameter processing
hkshell processes its arguments in a different way. Now it uses the standard
optparse module to parse the arguments. Documentation for hkshell has been
added, see the details there (both in the user documentation and in the
developer documentation of hkshell).

A utility function has been added to help defining user commands.