test_hklib: updated tests to suit thread links
Some changes were introduced due to changes indirectly introduced
during the change.  Namely, ˙˙Post˙˙'s ``htmlthreadfilename`` and
``htmlbasefilename`` check if the post in question is the root of a
thread. This, however, was not the case for one of the tests, which
called ``Generator.thread()`` on a non-root test post. Thus, this test
failed even when the necessary parameter for the thread link was added.
This test case turned out to be superfluous, as the next one is similar,
but starts off a root post.

Please note that though the tests now run OK, the new features are not
yet tested! See TODO.