GeneratorOptions: trycopycssfile option added
The new option is True by default. When the option is True, the Generator.gen_*
methods will copy the CSS file into the HTML directory if it isn't there yet.
This is handy for new users.

The solution is not yet perfect. This is due to our current limitation that the
GeneratorOptions.cssfile option is both the path of the CSS file and the CSS
file name in the "link rel=stylesheet" tag. Maybe we should separate these.