hkcustomlib.edit recognizes unmodified files
When hkcustomlib.edit is called, it is supposed to return True if the edited
file was changed and False otherwise. So far it always returned True. Now it
returns False if the file was not been changed.

The same was done in `hkrc_csabahoch.py`.

hkutils.file_to_string has a new optional parameter called `return_none`.
It specifies what to do if the file does not exist. If return_none is True,
None will be returned. Otherwise an exception will be raised.