hklib.__init__: better exception handling
If an exception is raised inside the parsing function, that exception is caught
and a new one is raised. So far the outer exception did not include the inner
exception, thus the exact description of the problem was lost. Now the outer
exception contains the inner one, so the information is kept.

This is not the best solution, though, because the resulted exception is ugly.
I think we need a HkParseException type which can be raised more conveniently
during parsing. This type could have a constructor that gets a file desciptor
and tries to read the file name from that.