hkshell.enew: new command for adding posts
`enew` creates and edits a post.

A temporary file with a post stub will be created and an editor will be opened
so that the user can edit it. After saving the file, it will be read as a post
file and the post created from it will be added to the post database. If the
post stub was not edited, nothing will happen.