hkshell.ls: new format
The `ls` command has a new format. It is easier to read. An example:

  >>> ls(ps(1802).exp())
  <1810> Should the "reviewed" tag be part of ...  Attis (2009.05.12. 16:55)
  <1658> Should the "reviewed" tag be part of ...  Csabi (2009.04.28. 23:29)
  <1689> Should the "reviewed" tag be part of ...  Csabi (2009.05.02. 15:17)
  <1802> Should the "reviewed" tag be part of ...  Csabi (2009.05.12. 07:17)
  <1797> Should the "reviewed" tag be part of ...  Attis (2009.05.11. 14:55)
  <1686> Should the "reviewed" tag be part of ...  Attis (2009.05.02. 12:42)
  <1831> Should the "reviewed" tag be part of ...  Csabi (2009.05.13. 08:43)
  <1798> Should the "reviewed" tag be part of ...  Csabi (2009.05.11. 16:07)
  <1577> Should the "reviewed" tag be part of ...  Attis (2009.04.17. 18:44)

Further improvements are needed. The biggest problem in my opinion is that
currently `ls` is not aware of the thread structure, dates, subjects, or really
anything. It just prints posts in the postset in an arbitrary order. It would
be very nice to have a similar output than in the HTML, using indentation.

A minor change: exit(0) has been changed to sys.exit(0) at one point of
hkshell.py. We use sys.exit everywhere, so I thought it is better to invoke the
same function everywhere to achieve the same goal.