Tutorial: half of the tutorial is almost ready
I added screenshots to the tutorial. When the tutorial is finished, I'll
redo the screenshots so that they will be consistent and they will have the
same window styles and anti-aliasing. These pictures were made on my UHU Linux
which had no anti-aliasing set. Since then I installed Kubuntu 9.4 with KDE 4
on that machine, so I would not be able to produce screenshots with the same
style even if I wanted to.

The documentation can be compiled without warning with Sphinx revision
9b28f4b1015c.