hkshell: Python interpreter invocation modified
The Python interpreter used to be invoked in a cumbersome way by Heapkeeper,
which turned out not to work well with Python 2.6. The old solution was that we
invoked `python` with a "-i" option, so after importing Heapkeeper modules, it
provided a Python shell to the user.

It turned out that there is a much better way to give Python shell to the
users: just call code.interact, which gives an interpreter to the user. hkshell
has been changed so that it uses code.interact now. A few other changes has
been made in hkshell, as well. The documentation has been changed accordingly,
and I also corrected a few minor things in hkshell.rst.

== The `hk` module and the `hkshell` sh script ==

A new module called `hk` has been introduced. This seemed necessary. First I
tried to open the interpreter from the hkshell module, but the problem was
that hkshell had to be fully imported. Thus hk imports hkshell, and the
interpreter is opened only afterwards. hk does not contain anything else.
Heapkeeper can be started now by just executing hk:

    $ python hk.py

This makes the `hkshell` sh script obsolete, so I modified it to print a
note about being deprecated. (There is no point in having both hk.py and
hkshell. The latter is easy to confuse with the hkshell Python module
(hkshell.py), and it makes it harder to write the documentation.)

=== Shell banner ===

A new option called hkshell.options.shell_banner has been introduced. The value
of this option is a string, which is printed when the Python shell starts.

=== CTRL-D ===

Heapkeeper's behaviour after the user presses CTRL-D has changed. Previosly we
could do nothing about it: we did not save anything. Now we ask the following
question from the user by default:

    Do you want to save the post database? [yes/no/cancel]

This behaviour can be overridden using the hkshell.options.save_on_ctrl_d
option.

== hkshell commands ==

The hkshell execution model became cleaner. I don't want to write about the
old model... The new model is that the hkshell.hkshell_commands dictionary
contains all hkshell commands. This dictionary can be given to the
code.interact function as the dictionary of local variables, which will result
in these commands being callable from the Python shell without having to
specify which module are they in.

This also means that from now, not every function and variable in the hkshell
module will be available from Heapkeeper's shell. Heapkeeper shell commands are
functions in the hkshell_commands dictionary, no matter whether they are
in the hkshell module or in some other module. Accessing everything in hkshell
can be turned on quite easily; maybe we should discuss whether we want it or
not.

=== hkshell.hkshell_cmd decorator ===

The hkshell.hkshell_cmd decorator can be used to create a command easily from a
function:

    @hkshell.hkshell_cmd()
    def my_command(...):
        ...

The decorator will add my_command to the user_commands dictionary. It also has
a few options; see its documentation for more information.

hkshell_cmd decorates now all functions in hkshell that are wanted to be
commands.

=== TouchedPostPrinterListener ===

This has been changed a bit. A bug has been corrected; now it prints out the
changed post in an ascending order. Also, there is no hardwired list with the
"touching" commands (=commands that activate TouchedPostPrinterListener); the
shell_cmd decorated can be used instead to declare a command to be "touching".

== Incompatibilities ==

- The `hkshell` sh script is deprecated. Use `python hk.py` instead.
- Decorator `hkshell.hkshell_events` has been renamed to `add_events`.
- Decorator `hkshell.shell_cmd` has been renamed to `hkshell_cmd`. Also, not it
  is not a decorator in the true sense but a decorator generator; which means
  that it needs an argument lis.