Post.date_str: uses customizable localtime_fun
Previously Post.date_str used time.localtime, which made it hard to test,
because the result was dependent on the time zone. Now it uses
hklib.localtime_fun, which points to time.localtime by default. This way the
default behaviour is the same, but the variable can be set for testing purposes
for example to time.gmtime, whose result is not dependent of time zone.