Module documentation and docstrings updated 1
Documentation has been updated so that it conforms to the form described in the
Documentation and Coding Conventions page (codingconventions.rst).

The module documentation (e.g. the <modname>.rst files) and the docstrings have
been updated for around half of the source code. All modules except for hklib
have been completed. In hklib, the update was started; the dosctrings for 540
lines out of 2640 has been update.