hklib: optimized IMAP downloading
The reason why the previous implementation was so slow is that it
performed many queries. 3 server queries were performed for each new
message, and 1 query for each message already present. Now only two
queries are performed:
- download all Message-Ids
- download all heads and bodies for new messages (if any)

This of course speeds up the process enormously, and also has the
advantage that we can tell the user the number of new messages before
downloading.