hklib.PostDB: walk_thread method created
The new method PostDB.walk_thread can iterate on threads in a more flexible
way than PostDB.iter_thread does. It has two advantages over iter_thread:

1. `walk_thread` yields each posts twice: when it enters the post's subthread
   and when it leaves it. This is very useful for example when generating the
   HTML pages: <div...> should be generated when entering into a post's
   subthread, and </div> when leaving it.

2. `walk_thread` returns the indentation level, as well.

`walk_thread` yields PostItem objects instead of Post objects. PostItem
objects point to one Post and store other information like the level of the
post in the thread structure. I use a PostItem instead of a tuple so that it
will be easy to add new information to the information yielded by
`walk_thread`.

Here follows the same example that is used in the test function.

The thread structure we walk on:

    0 <- 1 <- 2
      <- 3
    4

The post items yielded (indentation is there only to help the human reader):

    <PostItem: pos=begin, heapid='0', level=0>
      <PostItem: pos=begin, heapid='1', level=1>
        <PostItem: pos=begin, heapid='2', level=2>
        <PostItem: pos=end, heapid='2', level=2>
      <PostItem: pos=end, heapid='1', level=1>
      <PostItem: pos=begin, heapid='3', level=1>
      <PostItem: pos=end, heapid='3', level=1>
    <PostItem: pos=end, heapid='0', level=0>
    <PostItem: pos=begin, heapid='4', level=0>
    <PostItem: pos=end, heapid='4', level=0>