Post: "forward compatible" post file parsing
Previously, the post file parser (Post.create_header specifically) raised an
exception when it found an unknown attribute. Now it only prints a warning.

The reason for accepting unknown attributes are to be forward compatible.
Imagine that I create a new attribute called MyThing, and modify Hk so that it
uses that attribute:

    Author: Csabi
    MyThing: something
    MyThing: something else, too

    Hi.

All the other users that did not have the modified version of Hk would be
unable to use the heap that I modify with the modified Hk.

With the new behaviour, their Hk will just issue a Warning that they know
nothing about the 'MyThing' attribute. Even if they modify this post in Hk and
print it out, the MyThing attribute gets written out to the post file
unchanged.