Post.parse_header: "trailing \n appears" bug fixed
Bug's description
-----------------

When loading and saving a post with an attribute having multiple lines, an
additional newline appeared in the attribute.

Example:

Let's suppose this post file was read into a post:

    Author: me
    Subject: this is a very very very very
    <space>long subject
    Tags: example

When the post was saved (without modifications), the post file contained the
following:

    Author: me
    Subject: this is a very very very very
    <space>long subject
    <space>
    Tags: example

Investigation
-------------

hklib.Post.parse_header set post._header['Subject'] to be
'this is a very very very very\nlong subject\n' instead of
'this is a very very very very\nlong subject'.

Error
-----

There was an error in stipping the trailing newlines in Post.parse_header. This
line was in Post.parse_header:

    line.rstrip('\n')

This line does create a new string with trailing newlines removed, but does not
do anything with it. This new string is wanted to replace the `line` variable.
So the correct statement is this:

    line = line.rstrip('\n')

Note
----

A test case should be written that tests the parser against this error.