hklib, hkshell: dl() can return set of new posts
[post downloading]

Use ``dl(ps=True)`` if you want to use this, e.g.:

    >>> new = dl(ps=True)
    Reading settings...
    Connecting...
    [...]
    3 new messages downloaded.
    >>> ls(new)
    <101> Security breach at Bander's estate    Master Robot
    <102> Security breach examined      Security Robot
    <103> Urgent command    Master Robot

Without ``ps=True``, nothing is returned. This is needed because if the
user says a simple ``dl()``, the returned PostSet would be printed in
its string form, which is ugly if hundreds of new messages have been
downloaded.