Post.{write,write_str}: force_print pm added
[post editing]

The new `force_print` parameter is the set of attributes that should be printed
even if they are empty strings.

For example, normally, if the author is an empty string, it would not be
written into the post file. Sometimes, however, it is useful. E.g. consider
what happens when the user calls the enew command: a post file is shown to the
user with empty author and subject fields.