Html.post_summary: accepts a string as `tags`
[customizable post summary]

So far the `tags` argument has to be a list of tags, from which the
Html.post_summary function created a string by joining them and putting around
it a link to the post page.

Now it can be any HTML string. By overriding Generator.postitem_tags which
returns that HTML string, many nice things can be done now. Just three
interesting possibilities:

- We can display the differences in tags instead of all them:

  <0> Cats [funny, cute]    # has tags 'funny' and 'cute'
    <1> Cats [+small]       # has tags 'funny', 'cute' and 'small'
      <2> Cats [-funny]     # has tags 'cute' and 'small'
         <3> Cats           # has tags 'cute' and 'small'

- We can have colors in the tags. Important tags can be colored to be more
  noticeable.

- We can have links in the tags: for example each tag could link to an index
  page that contains all posts with that tag.