issue_tracker: added
[issue tracker]

`issue_tracker` is a new module that helps generating index pages that are made
for tracking Heapkeeper's issues. It also contains a few handy hkshell commands.

If you want to use the Issue Tracker, the simplest way is to append its
invocation to your `gen_indices` function in your hkrc::

    import hkshell
    import hkcustomlib

    def gen_indices(postdb):

        # Default index.html:
        hkcustomlib.gen_indices(postdb)

        # Own index pages
        # ...

        # Issue tracker
        genopts = hklib.GeneratorOptions()
        genopts.postdb = postdb
        generator = issue_tracker.create_generator(genopts)
        generator.gen_indices(genopts)

    hkshell.options.callbacks.gen_indices = gen_indices

This will generate two index pages when gi() or ga() is called:
- in issues_all.html, all issues are in one section
- in issues_sorted.html, the open and closed issues have separate sections

When Heapkeeper will have a better support for modules defined in a heap's
directory, maybe this module will be moved into the Hh's repository, together
with the CSS file (issues.css). Or maybe not.