test_hkutils: rearranging tests
[tidying]

This commit is a refactoring.

The test cases have been reordered so that they have the same order as the real
functions in hkutils. Their tester class has been renamed from TestUtilities to
TestMisc, since the functions tested there are referred to as "misc" functions
both in the source code of hktuils and in hkutils.rst.