hklib.PostItem: copies and prints unknown attribs
[new generator]

PostItem has currently three data attributes that is knows about. But since it
is for describing how a post should be printed, other attributes may be added
to a PostItem object. These attributes are now copied when the PostItem object
is copied, and they are printed when the object is converted to a string.