hklib: stars in flat sections bug fixed
[bugfix]

Code present in both `Generator.postitem_subject()` and
`Generator.postitem_tags()` decide whether the "star" would be used
instead of these fields. However, these latter functions do not care
whether the PostItem is flat or not, which results in the generation of
inappropriate stars.

I created a new `Generator.postitem_is_star_needed()` to check whether
a star is needed.  It is written to be generic, so if someone wants to
include authors in the "starring" process, it can be used for that, too.

This new function takes an argument, `param`, as the name of the
`Post` attribute to examine. Thus, if you choose to use it in
`Generator.postitem_subject()`, you can automatically use the option
`shortauthor` to enable or disable abbreviation of the author field.