PostDB.walk_thread: yield_main param added
[new generator]

The PostDB.walk_thread now has a yield_main parameter. It is False by default,
in which case the function uses the old behaviour.

If `yield_main` is True, an additional type of post item is returned, which is
called 'main'. (The "type" of the post item is called its "position").

An example of the yielded post items follows. The first word in the angle
brackets is the position of the post item. The second one is the heapid of the
post it points to. The indentation shows the `level` attribute of the post
item. The thread structure:

    0 <- 1 <- 2
      <- 3

The yielded post items:

    <begin 0>
    <main 0>
      <begin 1>
      <main 1>
        <begin 2>
        <main 2>
        <end 2>
      <end 1>
      <begin 3>
      <main 3>
      <end 3>
    <end 0>

The "main" postitem always follows the 'begin' postitem with the same post and
level. So the reader could ask: why do we need a separate "main" postitem?

The answer is that it is much easier to write general customization functions
in the new Generator if we have a separate "main" post item. Consider that we
want to see the summaries of posts with even heapids in red. Having the "main"
post item, we can write a post item modifier that will enclose all "main" post
items with red style:

    <begin 0>
    <red><main 0></red>
      <begin 1>
      <main 1>
        <begin 2>
        <red><main 2></red>
        <end 2>
      <end 1>
      <begin 3>
      <main 3>
      <end 3>
    <end 0>

An (untested) code follows that achieves this using the new Generator that is
to be commited shortly. Things that has to be known about the new Generator:
- It can print the given list of post items. That's why the following function
  will manipulate such a list.
- The `print_fun` data (!) attribute of the post items define how they should
  be printed. The print_fun receives a post item as an argument and returns an
  HtmlText (which is a TextStruct that contains HTML text).

The code:

    def red(xpostitems):
        """Prints xpostitems, but with the summaries of posts with
        even heapids in red.

        **Argument:**

        - `xpostitems` (iterable(|PostItem|))
        """

        def enclose_even_in_red(postitem):
            """Sets the print_fun attribute of the post item to print the post
            summary in red if the heapid of the post is even. In the end it
            returns the post item."""

            if ((postitem.pos == 'main') and # we want to enclose only the
                                             # "main" postitems
                (postitem.post.heapid() % 2 == 0)): # and the ones with even
                                                    # heapid

                old_print_fun = postitem.print_fun
                def new_print_fun(postitem):
                    """This function returns a HtmlText that is the same as the
                    one returned by the original printer function, except that
                    it is enclosed between "red" tags."""

                    return ('<red>', old_print_fun(postitem), '</red>')

                # We set the red-encloser function to be the printer function
                # of the post item.
                postitem.print_fun = new_print_fun

            return postitem

        return itertools.imap(enclose_even_in_red, xpostitems)