hkgen: new Generator class added
[new generator]

A new Generator class has been added to the (also new) hkgen module. The old
Generator will be removed at some point.

Not all methods of the new Generator is tested, but everything is documented.
The issue tracker is neither tested nor documented.

I will write a tutorial about the concepts of the new Generator. Until then,
probably the easiest way to learn it is to see how the issue tracker uses it.

I will spell out the main concept though. The main idea is that the post items
know how to print them, and we manipulate sequences of post items to print them
in customized ways. See the issue tracker for examples.

I haven't made measurements, but it seems to be faster than the old Generator,
thanks to using TextStruct objects instead of strings.

== Try it out ==

Try this to generate the new Generator:

    $ python hk.py --hkrc=hkrc_csabahoch 'gi();q()'

The following pages will be generated:
- index.html: main index page
- issues_all.html: all issues
- issues_sorted.html: all issues, but they are in different sections
- thread_<heapid>.html: thread pages

Don't forget to also move the heapindex.css and issues.css files to the HTML
directory.