Generator: all columns for flat postitems
[new generator]

Previously, the HTML output for flat postitems did not contain all fields. For
example if one post items did not print the "date" field, that row of the HTML
table had less column than the other ones.

A simplied example follows. The first post item does print a date, the
second one does not.

    <table>
    <tr><td class="author">Susan</td><td class="date">2009-01-01</td></tr>
    <tr><td class="author">Peter</td></tr>
    </table>

Now the output will look like this:

    <table>
    <tr><td class="author">Susan</td><td class="date">2009-01-01</td></tr>
    <tr><td class="author">Peter</td><td class="date"></td></tr>
    </table>

Also, previously, if the post had an empty body, or if the body printing was
turned off, the resulted HTML was the same. Now it is not the same, it will
generate a "body" element in the first case. (With no text inside, of course.)