Generator.reverse_threads added
[new generator]

The new method Generator.reverse_threads can reverse the threads in a sequence
of post sets.

You can have a look at the change in hkrc_csabahoch.print_main_index_page and
compare it to Generator.print_main_index_page to understand how to use it.

Also, the issue tracker uses the new method, so it shows the newest thread at
the top.