hklib: new mechanism to find next heapid w/ prefix
[next_heapid improvement]

Now results are consistent for empty and non-empty prefixes: in both
cases, the next heapid is defined as the one with the smallest number
that is larger than all numbers already present in heapids with a
matching prefix.

Also rewrote code that sets cached value for next heapid in
`PostDB._load_from_disk()`. Reloading the posts from disk resets the
cache.