hklib: next_heapid's caching extended for prefixes
[next_heapid improvement]

Previously the next heapid for the empty prefix was cached, it was
always recomputed for all others. Now the next heapid is cached for
all prefixes once it is computed.

The empty prefix remains special in that it gets cached during the
loading of the post database, while other prefixes get cached only
when first used.

Also added a test case for `PostDB.next_heapid`'s new features.