hkgen: copying CSS and other files in a better way
[new generator]

Copying non-generated files to the HTML directory (CSS files, images, etc.)
works now in the following way. The GeneratorOptions object has a files_to_copy
attribute that is a list. This should contain the names of the files that
should be copied into the HTML directory. Heapkeeper will try to copy these
files first from the directory of the post database, then from the current
directory (which should be Heapkeeper's main directory).

For example, in case of hkgen.Generator, `files_to_copy` is initialized as
follows:

    self.options.files_to_copy = ['heapindex.css', 'thread.png']

Since the pages generated by the issue tracker also require 'issues.css', the
issue tracker appends that to the list of files to be copied:

    self.options.files_to_copy.append('issues.css')