issue_tracker: showing meta info
[issue tracker]

The issue tracker now shows information like "effort", "version", "priority"
and "assigned to", which are read from the post bodies. The issue tracker page
shows only the value of the information (e.g. "1") at first and shows the type
of the information (e.g. "effort") only when the mouse is placed over the
value.

It would be good to improve the Generator so that it is not necessary to copy
the print_postitem_main function if we want to do something like this. But for
now, I copied that function into issue_tracker.py and modified it.