hkweb: added
[hkweb]

A dynamic web output is added in the separate file "hkweb.py".

hkweb is started from within hkshell like this:

    >>> import hkweb
    >>> hkweb.start()

Now open http://localhost:8080/ in your browser to view the heap's
index. If you change the post database in hkshell, the changes are
reflected in the browser whenever you refresh. This allows for a
faster view-edit-check cycle when tidying up the heap.