Generator: list of post summary fields extraced
[new generator]

== get_postsummary_fields_* functions ==

The list of post summary fields have been extracted into two functions [1]. The
functions that use this information [2] are simpler now, thanks to this
refactoring.

A consequence of this refactoring is that if someone wants to add a new field to
the post summary, only the simple get_postsummary_fields_* methods have to be
overridden instead of the more complicated print_postitem_*. (Which had been
even more complicated before this refactoring.)

This commit is not purely a refactoring because printing the flat posts is a bit
different from previously. (The details are not important here.)

== Two-level post summary field handling ==

Another change which makes it easier to modify the Generator's behaviour is the
two-level handling of post summary fields. Let's have a look at how the author
of the post item is printed:

    class Generator:

        ...

        def print_postitem_author_core(self, postitem):
            """Prints the core of the author of the post item.

            **Argument:**

            - `postitem` (|PostItem|)

            **Returns:** |HtmlText|
            """

            return self.escape(postitem.post.author())

        # TODO: test
        def print_postitem_author(self, postitem):
            """Prints the author of the post item.

            **Argument:**

            - `postitem` (|PostItem|)

            **Returns:** |HtmlText|
            """

            return self.enclose(
                       self.print_postitem_author_core(postitem),
                       class_='author',
                       skip_empty=True)

If a we want to highlight a specific author in our own Generator (which is a
subclass of hkgen.Generator), it is easy:

    class MyGenerator(hkgen.Generator):

        ...

        def print_postitem_author_core(self, postitem):

            # We get the default author core, which is an HTML text
            author_core = \
                hkgen.Generator.print_postitem_author_core(self, postitem)

            # We augment it
            if postitem.post.author() == 'Susan':
                author_core = self.enclose(author_core, class_='red')

            # We return the augmented version
            return author_core

This way both hkgen.Generator.print_postitem_author_core and
hkgen.Generator.print_postitem_author will be called, we just inserted our code
between them.

Of course something like this has to be put into the CSS file:

    .red
    {
        color: red;
    }

[1] hkgen.Generator.get_postsummary_fields_{main,flat}
[2] hkgen.Generator.print_postitem_{main,flat}