hkshell: grep added
[post search]

grep functionality has been added to hkshell. This is not the final version of
grep, because it should be implemented in hklib as a PostSet method. That would
be more convenient to use.

Examples of how to use the current implementation:

Posts that contain "hh": [1]

    >>> grep("master branch")
    PostSet([<post '3059'>, <post '2622'>])

Posts that contain "hh" and "master branch":

    >>> grep("master branch", grep("hh"))
    PostSet([<post '3059'>])

grep can be used together with gen_page, which generates an HTML page:

    >>> gen_page(grep("master branch", grep("hh")))

This will generate a page that shows that posts that contain both "master
branch" and "hh". The posts will be embedded in their thread structure; i.e.
their thread mates that do not contain the search expressions will be shown as
inactive.

[1] The real result is larger on our heap.