EmailDownloader: works with multipart messages
[email downloader]

Not only are there multipart messages when sending attachments, but the recent
text-only emails were multipart messages themselves, in the following way:

    - multipart/mixed container
        1. text/plain: email body
        2. text/plain: GMail banner

When sending HTML emails, the following seemed to be the message structure:

    - multipart/mixed container
        1. multipart/alternative container
            a. text/plain: email body in text
            b. text/html: email body in HTML
        2. text/html: GMail banner

So I implemented an algorithm in EmailDownloader.parse_email that finds the
first non-multipart message in this tree of messages and uses that.
EmailDownloader will print warnings if it meets some unforeseen content types.

EmailDownloader.parse_email have been split into two functions so that it will
be more testable.