hkutils: pylint warnings fixed
[pylint]

A little code has been changed according to the warnings printed by the pylint
static code analyzer (http://www.logilab.org/857).

There are still some pylint warnings. Now pylint gives no warning if it is
invoked in the following way:

    $ pylint --disable-msg=C0103,W0603 hkutils.py

pylint seems to be the most convenient when I called it in the following way:

    $ pylint --output=colorized \
             --reports=n \
             --include-ids=y \
             --disable-msg=C0103,W0603 hkutils.py