hkgen: '../' prepended to the links
[several heaps]

The new directory structure followed by introducing several heaps looked like
this:

    - heapindex.css
    - index.html
    - myheap
        - thread_0.html
        - thread_4.html

E.g. thread_0.html contained a reference to 'heapindex.css', which did not
work, and had to be replaced with '../heapindex.css'. But then also index.html
had to be copied into a directory so that it also can refer to
'../heapindex.css'.

So the new directory structure looks like this:

    - heapindex.css
    - index
        - index.html
    - myheap
        - thread_0.html
        - thread_4.html

Links had the same problem as the CSS files. Heapkeeper can generate links to
posts easily only if all HTML files are on the same level. So now all links in
the index pages and in the thread pages look like this:

    '../<heap id>/<html file>#<location>'