hk*: several heaps
[several heaps]

This commit introduces support for managing several heaps by Heapkeeper. This
means that a running instance of Heapkeeper can be aware of several heaps, and
handles them as one database.

== Configuration file format ==

The configuration file changed, but Heapkeeper remained backward compatible, so
it works with the old configuration file, as well.

An example of the old configuration file format:

    [paths]
    mail=/home/me/heaps/ums_posts
    html=/home/me/heaps/ums_html

An example of the new configuration file format:

    [paths]
    heaps=ums:/home/me/heaps/ums_posts;sandbox:/home/me/heaps/sandbox
    html=/home/me/heaps/ums_html

Heapkeeper remainging backward compatible means that if the config file
contains "mail=<heap_dir>" but does not contain "heaps" attribute, it is
equivalent to saying "heaps=:<heap_dir>", which means that there is one heap,
whose id is the empty string, and which is stored in heap_dir. The generated
HTML pages will be different from the HTML pages generated by the old
Heapkeeper, since the new pages will display the fact that the posts are
contained by the heap whose name is the empty string.

You can even use several heaps and yet make your config file work with the old
Heapkeeper by using both "mail" and "heaps" attributes:

    [paths]
    mail=/home/me/heaps/ums_posts
    heaps=ums:/home/me/heaps/ums_posts;sandbox:/home/me/heaps/sandbox
    html=/home/me/heaps/ums_html

Of course, the old Heapkeeper will be able to manage only one heap. The old
Heapkeeper will use only the "mail" attribute, while the new Heapkeeper will
use the "heaps" attribute.

== Post id, heap id and post index ==

We use now the concepts "post id", "heap id" and "post index" instead of the
old "heap id".

The old heap id was a string that identified a post (within the only heap).

In the new system, the heap id is the identifier (the name) of a heap. The post
name is the identifier of a post within a heap. (I.e. several posts may have
the same post index, but in different heaps.) The post id is a pair that
consists of a heap id and a post index: so it unambiguously identifies a post.
(Unambiguously, that is, within a configuration of a Heapkeeper: it is not as
unique as e.g. an URL or a Message id which can unambiguously identify
something within the whole world.)

== Automatic test and documentation improvements ==

Automatic tests have of course been updated. Tests in test_hklib were also
updated to our latest automatic test conventions. The same if true for module
documentation and docstrings.

== Plans ==

The killer feature of having several heaps would be a command to move a post
from one heap into another. That has not been implemented yet.